end ferst vertion of matrix lib
